further cleanup as we gain better understanding of the library we are updating comments to explain the commands of the ILI9486 TFT LCD Single Chip Driver.
One issue is the pixel addressing for accessing the display frame buffer
in order to do drawing. we have a function that sets the area to be
modified and then we set the RGB565 color value for each pixel in the
area. however addressing in the frame buffer is zero based and the range
is inclusive of the ending offset. so a region from point 1,0 to a point
of 1,3 is 4 pixels (1,0, 1,1, 1,2, 1,3) so if we are painting the pixels
in a 8x8 character, we need to paint from x0 to x0 + 7 for each row.

just a note that we need to check this addressing of regions carefully.