Added two new functions to TFTLCD class to perform touch hit calculation and check.  - bool TFTLCD::touch_circle()  - bool TFTLCD::touch_area()
added two new functions to Font class to provide text string sizes in
pixels for width of string and height of string.
 - Font::size_width()
 - Font::size_height()

Simplified the is_touch() function of the Button class by using our new
TFTLCD::touch_area() function along with a modified pendown() function
that now always returns boolean true. Got rid of the text string width
and height calculations in Button::drawButton() and using
Font::size_width() and Font:size_height() for positioning the text in
the button area.

what would be neat is to allow any geometric shape (rectangle, circle,
etc.) to be used as a button.

each Font object has its own copy of the FontTable flags. when the font
is created it captures a copy of the global FontTable flags but then a
person can modify the font specific FontTable flags. This is much more
flexible when mixing DoubleHigh and DoubleWide to modify the font
appearance. since Button::drawButton() uses a Font object for writing
the button caption text, we can now modify the text on a button by
button basis.