rewrote part of the Button class so that we now have a shared data area with a new class ButtonShared which saves some 11 bytes per button after the first button of the group sharing the data area.
Also changes to the TFTLCD class to provide for a point data structure,
TftPos, and a rectangle data structure, TftRect, to provide more
abstract graphical data items to make the job of using the KeDei TFT
library for a GUI more intuitive.

updated the scanner scale application for Arduino to use the new
functionality.